+++ modules/alloy/src/test/ts/browser/ui/form/HtmlSelectTest.ts
@@ -1,52 +1,91 @@
-import { ApproxStructure, Assertions, Logger, Pipeline, Step } from '@ephox/agar';
-import { UnitTest } from '@ephox/bedrock-client';
+import { ApproxStructure, Assertions } from '@ephox/agar';
+import { Assert, context, describe, it } from '@ephox/bedrock-client';
 
+import { Representing } from 'ephox/alloy/api/behaviour/Representing';
+import { AlloyComponent } from 'ephox/alloy/api/component/ComponentApi';
 import * as GuiFactory from 'ephox/alloy/api/component/GuiFactory';
 import * as GuiSetup from 'ephox/alloy/api/testhelpers/GuiSetup';
 import { HtmlSelect } from 'ephox/alloy/api/ui/HtmlSelect';
-import * as RepresentPipes from 'ephox/alloy/test/behaviour/RepresentPipes';
 
-UnitTest.asynctest('HtmlSelectTest', (success, failure) => {
+describe('HtmlSelectTest', () => {
+  const simpleAssert = (label: string, expected: any, component: AlloyComponent) =>
+    Assert.eq(label, expected, Representing.getValue(component));
 
-  const sNoInitialValue = Step.async((next, die) => {
-    GuiSetup.setup((_store, _doc, _body) => GuiFactory.build(
-      HtmlSelect.sketch({
-        dom: { }, // is always a select
+  context('no initial value', () => {
+    const store = GuiSetup.bddSetup(
+      () => GuiFactory.build(
+        HtmlSelect.sketch({
+          dom: {}, // is always a select
 
-        options: [
-          { value: 'alpha', text: 'Alpha' },
-          { value: 'beta', text: 'Beta' },
-          { value: 'gamma', text: 'Gamma' }
-        ]
-      })
-    ), (_doc, _body, _gui, component, _store) => [
-      RepresentPipes.sAssertValue('Checking initial value', 'alpha', component),
-      RepresentPipes.sSetValue(component, 'beta'),
-      RepresentPipes.sAssertValue('Checking value after valid set', 'beta', component),
-      RepresentPipes.sSetValue(component, 'delta'),
-      RepresentPipes.sAssertValue('Checking value after invalid set (should still be on beta)', 'beta', component),
-      RepresentPipes.sSetValue(component, 'gamma'),
-      RepresentPipes.sAssertValue('Checking value after valid set (should now be gamma)', 'gamma', component)
-    ], next, die);
+          options: [
+            { value: 'alpha', text: 'Alpha' },
+            { value: 'beta', text: 'Beta' },
+            { value: 'gamma', text: 'Gamma' }
+          ]
+        })
+      )
+    );
+
+    it('TBA: Run tests', () => {
+      const component = store.component();
+
+      simpleAssert('Checking initial value', 'alpha', component);
+      Representing.setValue(component, 'beta');
+      simpleAssert('Checking value after valid set', 'beta', component);
+      Representing.setValue(component, 'delta');
+      simpleAssert('Checking value after invalid set (should still be on beta)', 'beta', component);
+      Representing.setValue(component, 'gamma');
+      simpleAssert('Checking value after valid set (should now be gamma)', 'gamma', component);
+    });
+
+    it('TINY-9679: The options are changed and the selection reset', () => {
+      const component = store.component();
+
+      Assertions.assertStructure(
+        'Checking initial structure',
+        ApproxStructure.build((s, str) => s.element('select', {
+          children: [
+            s.element('option', { value: str.is('alpha'), html: str.is('Alpha') }),
+            s.element('option', { value: str.is('beta'), html: str.is('Beta') }),
+            s.element('option', { value: str.is('gamma'), html: str.is('Gamma') })
+          ]
+        })),
+        component.element
+      );
+
+      // component is reused from previous test, and starts on gamma.
+      simpleAssert('Checking initial value', 'gamma', component);
+      Representing.setValue(component, '');
+      simpleAssert('Checking initial value', 'gamma', component);
+      component.element.dom.value = null;
+      Representing.setValue(component, '');
+      simpleAssert('Checking value after valid set (should now be alpha)', 'alpha', component);
+    });
   });
 
-  const sHasInitialValue = Step.async((next, die) => {
-    GuiSetup.setup((_store, _doc, _body) => GuiFactory.build(
-      HtmlSelect.sketch({
-        dom: { }, // is always a select
-        selectAttributes: {
-          size: 10
-        },
-        selectClasses: [ 'my-test-select' ],
-        data: 'gamma',
-        options: [
-          { value: 'alpha', text: 'Alpha' },
-          { value: 'beta', text: 'Beta' },
-          { value: 'gamma', text: 'Gamma' }
-        ]
-      })
-    ), (_doc, _body, _gui, component, _store) => [
-      Assertions.sAssertStructure(
+  context('Has initial value gamma', () => {
+    const store = GuiSetup.bddSetup(
+      () => GuiFactory.build(
+        HtmlSelect.sketch({
+          dom: {}, // is always a select
+          selectAttributes: {
+            size: 10
+          },
+          selectClasses: [ 'my-test-select' ],
+          data: 'gamma',
+          options: [
+            { value: 'alpha', text: 'Alpha' },
+            { value: 'beta', text: 'Beta' },
+            { value: 'gamma', text: 'Gamma' }
+          ]
+        })
+      )
+    );
+
+    it('TBA: Run tests', () => {
+      const component = store.component();
+
+      Assertions.assertStructure(
         'Checking initial structure',
         ApproxStructure.build((s, str, arr) => s.element('select', {
           attrs: {
@@ -60,22 +99,17 @@ UnitTest.asynctest('HtmlSelectTest', (success, failure) => {
           ]
         })),
         component.element
-      ),
+      );
 
-      RepresentPipes.sAssertValue('Checking initial value', 'gamma', component),
-      RepresentPipes.sSetValue(component, 'beta'),
-      RepresentPipes.sAssertValue('Checking value after valid set', 'beta', component),
-      RepresentPipes.sSetValue(component, 'delta'),
-      RepresentPipes.sAssertValue('Checking value after invalid set (should still be on beta)', 'beta', component),
-      RepresentPipes.sSetValue(component, 'gamma'),
-      RepresentPipes.sAssertValue('Checking value after valid set (should now be gamma)', 'gamma', component),
-      RepresentPipes.sSetValue(component, 'alpha'),
-      RepresentPipes.sAssertValue('Checking value after valid set (should now be alpha)', 'alpha', component)
-    ], next, die);
+      simpleAssert('Checking initial value', 'gamma', component);
+      Representing.setValue(component, 'beta');
+      simpleAssert('Checking value after valid set', 'beta', component);
+      Representing.setValue(component, 'delta');
+      simpleAssert('Checking value after invalid set (should still be on beta)', 'beta', component);
+      Representing.setValue(component, 'gamma');
+      simpleAssert('Checking value after valid set (should now be gamma)', 'gamma', component);
+      Representing.setValue(component, 'alpha');
+      simpleAssert('Checking value after valid set (should now be alpha)', 'alpha', component);
+    });
   });
-
-  Pipeline.async({}, [
-    Logger.t('Scenario: no initial value', sNoInitialValue),
-    Logger.t('Scenario: has initial value gamma', sHasInitialValue)
-  ], success, failure);
 });