+++ modules/alloy/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- Select would in some situations not have a selected value when expected to. #TINY-9679
+
 ## 12.1.0 - 2023-03-15
 
 ### Added